refactor: isolate ComboBox stories
